refact process operation
