+++ src/cli/node/config/config.js
@@ -15,7 +15,8 @@ const { fromSecretKey } = require('@arcblock/forge-wallet');
 const { bytesToHex, hexToBytes, isHexStrict } = require('@arcblock/forge-util');
 const GraphQLClient = require('@arcblock/graphql-client');
 const { symbols, hr, pretty } = require('core/ui');
-const { config, requiredDirs, webUrl, findServicePid, ensureConfigComment } = require('core/env');
+const { config, requiredDirs, webUrl, ensureConfigComment } = require('core/env');
+const { findServicePid } = require('forge-process');
 
 function getModeratorSecretKey() {
   const sk = process.env.FORGE_MODERATOR_SK;
