+++ src/common.js
@@ -1,8 +1,18 @@
 const shell = require('shelljs');
-
+const prettyMilliseconds = require('pretty-ms');
 const { symbols } = require('core/ui');
 
+function prettyTime(ms) {
+  let result = prettyMilliseconds(ms, { compact: true });
+  if (result.startsWith('~')) {
+    result = result.slice(1);
+  }
+
+  return result;
+}
+
 const Common = {
+  prettyTime,
   print: content => {
     shell.echo(`${content}`);
   },
